add automapper maps for timebank to Timebank refactor tests to verify automapper configs remove EntityTranslator class
