+++ solution/Timebanks.NZ.DAL.MySql/MySqlPocoToBEProfile.cs
@@ -16,9 +16,12 @@ namespace Timebanks.NZ.DAL.MySql
     {
         protected override void Configure()
         {
+            // For some reason automapper does not map properties back the other way but it should...
             Mapper.CreateMap<member, User>()
                 .ForMember(x => x.IdMember, opt => opt.ResolveUsing(member => member.id_member))
                 .ForMember(x => x.IdTimebank, opt => opt.ResolveUsing(member => member.id_timebank));
+
+            Mapper.CreateMap<timebank, Timebank>();
 
             SourceMemberNamingConvention = new LowerUnderscoreNamingConvention();
             DestinationMemberNamingConvention = new PascalCaseNamingConvention();
@@ -33,7 +36,7 @@ namespace Timebanks.NZ.DAL.MySql
     {
         protected override void Configure()
         {          
-            // Ignore navigation properties
+            // Ignore EF navigation properties
             Mapper.CreateMap<User, member>()
                 .ForMember(x => x.offer_need, opt => opt.Ignore())
                 .ForMember(x => x.timebank, opt => opt.Ignore())
@@ -41,7 +44,11 @@ namespace Timebanks.NZ.DAL.MySql
                 .ForMember(x => x.member_permission, opt => opt.Ignore())
                 .ForMember(x => x.trades, opt => opt.Ignore())
                 .ForMember(x => x.trades1, opt => opt.Ignore());
-                
+
+            Mapper.CreateMap<Timebank, timebank>()
+                .ForMember(x => x.offer_need, opt => opt.Ignore())
+                .ForMember(x => x.members, opt => opt.Ignore());
+
             SourceMemberNamingConvention = new PascalCaseNamingConvention();
             DestinationMemberNamingConvention = new LowerUnderscoreNamingConvention();
         }
