+++ solution/TimebanksNZ.Tests/TimebanksNZ.Tests.csproj
@@ -122,7 +122,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="DAL\MySql\EntityTranslatorTests.cs" />
+    <Compile Include="DAL\MySql\AutomapperTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Controllers\HomeControllerTest.cs" />
     <Compile Include="UserRepositoryTests.cs" />
